% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spTempApprox.R
\name{spTempApprox}
\alias{spTempApprox}
\title{MCMC Algorithm For Estimation Of Local Model Through Scalable (Approximate)
Estimation}
\usage{
spTempApprox(
  y,
  X,
  c_init,
  beta_init,
  zeta_init,
  nusq_init,
  l_init,
  prior.mu,
  Spat.cov,
  Temp.cov,
  v0,
  a1,
  a2,
  q,
  Nmcmc,
  ind
)
}
\arguments{
\item{y}{binary response n values}

\item{X}{data - dimension \eqn{n}x\eqn{L}x\eqn{\tau}}

\item{c_init}{initial value for \eqn{c} - dimension \eqn{\tau}}

\item{beta_init}{initial values for \eqn{\beta} - dimension \eqn{L}x\eqn{\tau}}

\item{zeta_init}{initial values for \eqn{\zeta} - dimension \eqn{L}x\eqn{\tau}}

\item{nusq_init}{initial values for \eqn{\nu^2} - dimension \eqn{L}x\eqn{\tau}}

\item{l_init}{initial value for \eqn{\lambda} - dimension \eqn{L\tau}}

\item{prior.mu}{prior mean for \eqn{\lambda} - dimension \eqn{L\tau}}

\item{Spat.cov}{spatial covariance matrix  - dimension \eqn{L}x\eqn{L}}

\item{Temp.cov}{temporal covariance matrix  - dimension \eqn{\tau}x\eqn{\tau}}

\item{v0}{hyperparameter \eqn{v_0}}

\item{a1}{hyperparameter \eqn{a_1}}

\item{a2}{hyperparameter \eqn{a_2}}

\item{q}{hyperparameter \eqn{q}}

\item{Nmcmc}{number of MCMC samples to generate}

\item{ind}{indices of MCMC samples to use}
}
\value{
\itemize{
\item{\code{b}} {MCMC samples for \eqn{\beta} - dimension (size of
                 \code{ind})x\eqn{L}x\eqn{\tau}}
\item{\code{zeta}} {MCMC samples for \eqn{\zeta} - dimension (size of
                 \code{ind})x\eqn{p}x\eqn{\tau}}
\item{\code{l}} {MCMC samples for \eqn{\lambda} - dimension (size of
                 \code{ind})x\eqn{L\tau}}
\item{\code{c}} {MCMC samples for \eqn{c} - dimension (size of
                 \code{ind})x\eqn{L\tau}}
\item{\code{t0}} {Samples for \eqn{t0} - dimension (size of
                 \code{ind})}
}
}
\description{
Gibbs sampling, and Metropolis-Hastings within Gibbs sampling for a local
model through scalable (approximate) estimation.
}
