% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spTempBayes.R
\name{spTempBayes}
\alias{spTempBayes}
\title{Bayesian Estimation (Through Exact/Approximate Estimation) And Prediction}
\usage{
spTempBayes(
  y,
  X,
  c_init = rep(1, tau),
  beta_init = NULL,
  zeta_init = NULL,
  nusq_init = NULL,
  l_init = NULL,
  prior.mu,
  Spat.cov,
  Temp.cov,
  v0,
  a1,
  a2,
  q,
  Nmcmc,
  burnin,
  X.test = NULL,
  type
)
}
\arguments{
\item{y}{binary response n values}

\item{X}{data - dimension \eqn{n}x\eqn{L}x\eqn{\tau}}

\item{c_init}{initial value for \eqn{c} - dimension \eqn{\tau} (defaults to 1)}

\item{beta_init}{initial values for \eqn{\beta} - dimension \eqn{L}x\eqn{\tau}
(defaults to 0)}

\item{zeta_init}{initial values for \eqn{\zeta} - dimension \eqn{L}x\eqn{\tau}
(defaults to randomly generated values from 1 or \eqn{v_0})}

\item{nusq_init}{initial values for \eqn{\nu^2} - dimension \eqn{L}x\eqn{\tau}
(defaults to 1)}

\item{l_init}{initial value for \eqn{\lambda} - dimension \eqn{L\tau}
(defaults to 0)}

\item{prior.mu}{prior mean for \eqn{\lambda} - dimension \eqn{L\tau}}

\item{Spat.cov}{spatial covariance matrix  - dimension \eqn{L}x\eqn{L}}

\item{Temp.cov}{temporal covariance matrix  - dimension \eqn{\tau}x\eqn{\tau}}

\item{v0}{hyperparameter \eqn{v_0}}

\item{a1}{hyperparameter \eqn{a_1}}

\item{a2}{hyperparameter \eqn{a_2}}

\item{q}{hyperparameter \eqn{q}}

\item{Nmcmc}{number of MCMC samples to generate}

\item{burnin}{number of samples to be dropped as burnin values}

\item{X.test}{Test data samples - dimension \eqn{m}x\eqn{L}x\eqn{\tau}}

\item{type}{Type of estimation (choose between 'exact' or 'approx')}
}
\value{
\itemize{
\item{\code{b.strucBayes}} {estimated coefficients \eqn{\beta} - dimension
                            \eqn{\tau}x\eqn{L}}
\item{\code{p.strucBayes.pred}} {estimated local prediction probabilities
                                 - dimension \eqn{m}x\eqn{\tau}}
\item{\code{y.strucBayes.pred}} {predicted response - length \eqn{m}}
\item{\code{run.time}} {total run time}
}
}
\description{
Bayesian variable selection for the EEG data. The estimation type (exact or
approx) could needs to be specified. Also performs prediction for a test data
set. (Warning: type='exact' will require huge amount of time to execute even
for moderate values of \eqn{L} or \eqn{\tau}.)
}
